pyproject.toml
[project]
name = "custom-resume"
version = "0.1.0"
description = "ATS-optimized resume generation system"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # AI
    "google-genai>=1.2.0",
    # PDF processing
    "pdfplumber>=0.11.0",
    "pdf2image>=1.17.0",
    # Pydantic
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Auth
    "pyjwt>=2.8.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    # Storage
    "google-cloud-storage>=2.16.0",
    # Payments
    "razorpay>=1.4.0",
    # Chat (ADK + JSON Patch)
    "google-adk>=1.0.0",
    "jsonpatch>=1.33",
    # Utils
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
